# docs/swagger.yaml
definitions:
  http_interfaces_authentication.AuthenticationResponse:
    properties:
      token:
        type: string
    type: object
  http_interfaces_authentication.CreateAuthenticationRequest:
    properties:
      email:
        type: string
      name:
        type: string
    required:
    - email
    - name
    type: object
  http_interfaces_customer.CreateCustomerRequest:
    properties:
      email:
        maxLength: 254
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
    required:
    - email
    - name
    type: object
  http_interfaces_customer.CustomerResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  http_interfaces_customer.UpdateCustomerRequest:
    properties:
      email:
        maxLength: 254
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
    type: object
  http_interfaces_product.CreateProductRequest:
    properties:
      description:
        maxLength: 500
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      price:
        minimum: 0
        type: number
    required:
    - name
    - price
    type: object
  http_interfaces_product.ProductResponse:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: number
    type: object
  http_interfaces_product.UpdateProductRequest:
    properties:
      description:
        maxLength: 500
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      price:
        minimum: 0
        type: number
    type: object
  http_interfaces_tab.CreateTabRequest:
    properties:
      email:
        maxLength: 254
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
    required:
    - email
    - name
    type: object
  http_interfaces_tab.TabResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  http_interfaces_tab.UpdateTabRequest:
    properties:
      email:
        maxLength: 254
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
    type: object
info:
  contact: {}
  description: API para favoritar produtos
  title: Favorite API
  version: "1.0"
paths:
  /authenticate:
    post:
      consumes:
      - application/json
      description: Creates a JWT token for the user
      parameters:
      - description: Authentication data
        in: body
        name: auth
        required: true
        schema:
          $ref: '#/definitions/http_interfaces_authentication.CreateAuthenticationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http_interfaces_authentication.AuthenticationResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create authentication token
      tags:
      - Authentication
  /customer:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated list of customers
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/http_interfaces_customer.CustomerResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all customers
      tags:
      - Customer
    post:
      consumes:
      - application/json
      description: Creates a favorite for a given customer and product
      parameters:
      - description: Customer data
        in: body
        name: favorite
        required: true
        schema:
          $ref: '#/definitions/http_interfaces_customer.CreateCustomerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http_interfaces_customer.CustomerResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add customer product
      tags:
      - Customer
  /customer/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a customer by their ID
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete customer
      tags:
      - Customer
    get:
      consumes:
      - application/json
      description: Retrieves a customer's information by their ID
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http_interfaces_customer.CustomerResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get customer by ID
      tags:
      - Customer
    put:
      consumes:
      - application/json
      description: Updates a customer's information by their ID
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated customer data
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/http_interfaces_customer.UpdateCustomerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http_interfaces_customer.CustomerResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update customer
      tags:
      - Customer
  /product:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated list of products
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/http_interfaces_product.ProductResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all products
      tags:
      - Product
  /product/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a product by their ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete product
      tags:
      - Product
    get:
      consumes:
      - application/json
      description: Retrieves a product's information by their ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http_interfaces_product.ProductResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get product by ID
      tags:
      - Product
    put:
      consumes:
      - application/json
      description: Updates a product's information by their ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated product data
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/http_interfaces_product.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http_interfaces_product.ProductResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update product
      tags:
      - Product
  /tab:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated list of tabs
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/http_interfaces_tab.TabResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all tabs
      tags:
      - Tab
    post:
      consumes:
      - application/json
      description: Creates a favorite for a given tab and product
      parameters:
      - description: Tab data
        in: body
        name: favorite
        required: true
        schema:
          $ref: '#/definitions/http_interfaces_tab.CreateTabRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http_interfaces_tab.TabResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add tab product
      tags:
      - Tab
  /tab/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a tab by their ID
      parameters:
      - description: Tab ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete tab
      tags:
      - Tab
    get:
      consumes:
      - application/json
      description: Retrieves a tab's information by their ID
      parameters:
      - description: Tab ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http_interfaces_tab.TabResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get tab by ID
      tags:
      - Tab
    put:
      consumes:
      - application/json
      description: Updates a tab's information by their ID
      parameters:
      - description: Tab ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated tab data
        in: body
        name: tab
        required: true
        schema:
          $ref: '#/definitions/http_interfaces_tab.UpdateTabRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http_interfaces_tab.TabResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update tab
      tags:
      - Tab
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"